Remove central error handling for better debugging
As far as I experience, the central exception handler obfuscates the
error traceback, as it just redisplays the main error message, but not
the stack trace.

As I need to see in the inspector console where an error comes from, I
removed the central error handling.

If that can be done in a way that doesn't take away the stack trace
information, or if I understood something wrong, feel free to reject
this.