Expose user name and admin state in auth module
Due to liqd/adhocracy3#1635, userData isn't submitted properly anymore.
Therefore we fetch the user again on policycompass side in order to
retrieve the username. This can be removed later.

In addition, we now set Auth.state.isAdmin if the user is an admin user
(i.e. belongs to the gods group and thus has the god role in Adhocracy).

Fixes policycompass/policycompass#121