Redirect to previous page after login / register
Fixes policycompass/policycompass#313.

This uses the `$locationChangeSuccess` event in order to get the
previous URL and gets the path by splitting the URL at `#!`. This isn't
ideal, as it requires the application to run in hashbang mode and would
break in html5 mode.

Alternatively, the `$routeChangeSuccess` event could be used, but this
would require to the actual path using based on the path templates (e.g.
`/datasets/:datasetId`) and parameters (e.g. `{'dataset': 15}`).

The same could also be done on logout, but I didn't do as it's less
relevant and not clear whether it's actually desired. More importantly,
this wouldn't work if a user visited a restricted page before logout.